TargetAndActualChart

+++ budgeteer-web-interface/src/main/java/org/wickedsource/budgeteer/web/usecase/dashboard/component/burnedbudgetchart/BurnedBudgetChartModel.java
@@ -27,7 +27,7 @@ public class BurnedBudgetChartModel extends LoadableDetachableModel<List<Double>
         return service.getBudgetBurnedInPreviousWeeks(userId, numberOfWeeks);
     }
 
-    public long getNumberOfWeeks() {
+    public int getNumberOfWeeks() {
         return this.numberOfWeeks;
     }
 }
